+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [hooks:init] Allow to customize help.
 - [child-processes] `DEBUG_CP` environment variable for debugging child processes.
 - Command `vtex lighthouse` for vtex internal Google's Lighthouse tool
+- [vtex promote] Save workspace promotion info in evolution manager.
 
 ### Changed
 - Update release notes message.